Start refactoring Elm
