+++ elm/lib/dependabot/elm/update_checker/cli_parser.rb
@@ -1,31 +1,29 @@
 # frozen_string_literal: true
 
 require "dependabot/utils/elm/version"
-require "dependabot/update_checkers/elm/elm_package"
+require "dependabot/elm/update_checker"
 
 module Dependabot
-  module UpdateCheckers
-    module Elm
-      class ElmPackage
-        class CliParser
-          INSTALL_DEPENDENCY_REGEX =
-            %r{([^\s]+\/[^\s]+)\s+(\d+\.\d+\.\d+)}.freeze
-          UPGRADE_DEPENDENCY_REGEX =
-            %r{([^\s]+\/[^\s]+) \(\d+\.\d+\.\d+ => (\d+\.\d+\.\d+)\)}.freeze
+  module Elm
+    class UpdateChecker
+      class CliParser
+        INSTALL_DEPENDENCY_REGEX =
+          %r{([^\s]+\/[^\s]+)\s+(\d+\.\d+\.\d+)}.freeze
+        UPGRADE_DEPENDENCY_REGEX =
+          %r{([^\s]+\/[^\s]+) \(\d+\.\d+\.\d+ => (\d+\.\d+\.\d+)\)}.freeze
 
-          def self.decode_install_preview(text)
-            installs = {}
+        def self.decode_install_preview(text)
+          installs = {}
 
-            # Parse new installs
-            text.scan(INSTALL_DEPENDENCY_REGEX).
-              each { |n, v| installs[n] = Utils::Elm::Version.new(v) }
+          # Parse new installs
+          text.scan(INSTALL_DEPENDENCY_REGEX).
+            each { |n, v| installs[n] = Utils::Elm::Version.new(v) }
 
-            # Parse upgrades
-            text.scan(UPGRADE_DEPENDENCY_REGEX).
-              each { |n, v| installs[n] = Utils::Elm::Version.new(v) }
+          # Parse upgrades
+          text.scan(UPGRADE_DEPENDENCY_REGEX).
+            each { |n, v| installs[n] = Utils::Elm::Version.new(v) }
 
-            installs
-          end
+          installs
         end
       end
     end
